% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{read_table}
\alias{read_table}
\title{Simple read table}
\usage{
read_table(conn_str, name)
}
\description{
Simple read table
}
\details{
This impl creates ad-hoc connection which has significant overhead
}
\examples{
# none so far
}
